refactor: move AppMercadoLivre to new folder
